requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
django-crispy-forms==1.7.2
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-pubsub==1.0.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
